requirements.txt
python-dotenv
requests
python-dateutil
Pillow
numpy
sty
PyYAML